phoenix/esgf/views/esgfsearch.py
from deform import Form
from pyramid.view import view_defaults

from phoenix.esgf.schema import ESGFSearchSchema
from phoenix.esgf.search import ESGFSearch


@view_defaults(permission='view', layout='default')
class ESGFSearchActions(object):
    def __init__(self, request):
        self.request = request
        self.session = request.session
        self.esgfsearch = ESGFSearch(request)

    def search_datasets(self):
        result = dict()
        result.update(self.esgfsearch.query_params())
        result.update(self.esgfsearch.search_datasets())
        result['form'] = Form(ESGFSearchSchema())
        result['quickview'] = True
        return result

    def search_items(self):
        return self.esgfsearch.search_items()
